requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_r-mutex=1.0.0=anacondar_1
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py312hd77b12b_7
ca-certificates=2024.3.11=haa95532_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
email-validator=2.1.1=pyhd8ed1ab_0
flask-login=0.6.3=pyhd8ed1ab_1
flask-sqlalchemy=3.0.3=pyhd8ed1ab_0
flask-wtf=1.2.1=pyhd8ed1ab_0
glib-tools=2.78.4=hd77b12b_0
gst-plugins-base=1.18.5=h9e645db_0
importlib-metadata=7.0.1=pyha770c72_0
intel-openmp=2024.1.0=h57928b3_965
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
m2-base=1.0.0=3
m2-bash=4.3.042=5
m2-bash-completion=2.3=2
m2-catgets=1.1=3
m2-coreutils=8.25=102
m2-dash=0.5.8=2
m2-file=5.25=2
m2-filesystem=2016.04=4
m2-findutils=4.6.0=2
m2-gawk=4.1.3=2
m2-gcc-libs=5.3.0=4
m2-gettext=0.19.7=4
m2-gmp=6.1.0=3
m2-grep=2.22=4
m2-gzip=1.7=2
m2-inetutils=1.9.2=2
m2-info=6.0=2
m2-less=481=2
m2-libasprintf=0.19.7=4
m2-libcatgets=1.1=3
m2-libcrypt=1.3=2
m2-libgettextpo=0.19.7=4
m2-libiconv=1.14=3
m2-libintl=0.19.7=4
m2-liblzma=5.2.2=2
m2-libpcre=8.38=2
m2-libreadline=6.3.008=8
m2-libutil-linux=2.26.2=2
m2-libxml2=2.9.2=3
m2-mintty=1!2.2.3=2
m2-mpfr=3.1.4=2
m2-msys2-launcher-git=0.3.28.860c495=2
m2-msys2-runtime=2.5.0.17080.65c939c=3
m2-ncurses=6.0.20160220=2
m2-sed=4.2.2=3
m2-tftp-hpa=5.2=2
m2-time=1.7=2
m2-ttyrec=1.0.8=2
m2-tzcode=2015.e=2
m2-util-linux=2.26.2=2
m2-which=2.21=3
m2-zlib=1.2.8=4
m2w64-bwidget=1.9.10=2
m2w64-bzip2=1.0.6=6
m2w64-expat=2.1.1=2
m2w64-fftw=3.3.4=6
m2w64-flac=1.3.1=3
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gettext=0.19.7=2
m2w64-gmp=6.1.0=2
m2w64-gsl=2.1=2
m2w64-libiconv=1.14=6
m2w64-libjpeg-turbo=1.4.2=3
m2w64-libogg=1.3.2=3
m2w64-libpng=1.6.21=2
m2w64-libsndfile=1.0.26=2
m2w64-libtiff=4.0.6=2
m2w64-libvorbis=1.3.5=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-libxml2=2.9.3=4
m2w64-mpfr=3.1.4=4
m2w64-pcre2=10.34=0
m2w64-speex=1.2rc2=3
m2w64-speexdsp=1.2rc3=3
m2w64-tcl=8.6.5=3
m2w64-tk=8.6.5=3
m2w64-tktable=2.10=5
m2w64-wineditline=2.101=5
m2w64-xz=5.2.2=2
m2w64-zlib=1.2.8=10
matplotlib-base=3.8.4=py312hc7c4135_0
msys2-conda-epoch=20160418=1
psycopg-c=3.1.18=py312hf4fdf6b_0
pyqt5-sip=12.13.0=py312h2bbff1b_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-docx=1.1.0=pyhd8ed1ab_0
python-dotenv=1.0.0=pyhd8ed1ab_1
python-tzdata=2023.3=pyhd3eb1b0_0
qt-main=5.15.2=he8e5bd7_7
r-base=4.1.3=hdca333a_12
typing-extensions=4.9.0=py312haa95532_1
